Documentation first pass
This is the first full pass at writing and updating the formatting of documentation. For all functions, existing documentation was moved to the beginning of the file, before the code began so it can be read by comment2docs.py. Other formatting changes were implemented. 'docs' folder and files within was created.