Updated Protocol and CS (#228)
* Update Protocol.m

Added an if statement to switch between instruction screens for the original and 2afc protocol.

* Update cs.m

Added a ~ in front of the license function, due to the license function outputting a 1 if the license is not there, rather than a 0, which was most likely expected.

* Update Protocol.m

Changed how the fixation screens are generated form the previous change. Also fixed a mistake where the default instructions were presented for the 2afc protocol.

* Added new fixation screens for 2afc

Added new fixation screens to give direction for the 2afc experiment.